Remove RevisionDiffController from FormDiff and FileStatusList (cont.)
