+++ GitCommands/Git/GitRevisionTester.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO.Abstractions;
 using System.Linq;
 
 namespace GitCommands.Git
@@ -7,11 +8,26 @@ namespace GitCommands.Git
     public interface IGitRevisionTester
     {
         bool IsFirstParent(GitRevision revision, IEnumerable<GitRevision> selectedItemParents);
+        bool LocalRevisionExists(IEnumerable<GitItemStatus> selectedItemsWithParent);
         bool Matches(GitRevision revision, string criteria);
     }
 
     public class GitRevisionTester : IGitRevisionTester
     {
+        private readonly IFullPathResolver _fullPathResolver;
+        private readonly IFileSystem _fileSystem;
+
+        public GitRevisionTester(IFullPathResolver fullPathResolver, IFileSystem fileSystem)
+        {
+            _fullPathResolver = fullPathResolver;
+            _fileSystem = fileSystem;
+        }
+
+        public GitRevisionTester(IFullPathResolver fullPathResolver)
+            : this(fullPathResolver, new FileSystem())
+        {
+        }
+
         public bool IsFirstParent(GitRevision revision, IEnumerable<GitRevision> selectedItemParents)
         {
             if (revision?.ParentGuids == null)
@@ -29,6 +45,33 @@ namespace GitCommands.Git
             }
 
             return true;
+        }
+
+        public bool LocalRevisionExists(IEnumerable<GitItemStatus> selectedItemsWithParent)
+        {
+            if (selectedItemsWithParent == null)
+            {
+                return false;
+            }
+
+            var items = selectedItemsWithParent as List<GitItemStatus> ?? selectedItemsWithParent.ToList();
+            bool localExists = items.Any(item => !item.IsTracked);
+            if (localExists)
+            {
+                return true;
+            }
+
+            // enable *<->Local items only when (any) local file exists
+            foreach (var item in items)
+            {
+                string filePath = _fullPathResolver.Resolve(item.Name);
+                if (_fileSystem.File.Exists(filePath))
+                {
+                    return true;
+                }
+            }
+
+            return false;
         }
 
         public bool Matches(GitRevision revision, string criteria)
